backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.1.0
sqlalchemy==2.0.34
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1